api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
scikit-learn==1.4.2
scipy==1.13.0
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
google-generativeai>=0.7.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pdfplumber>=0.10.0
python-multipart>=0.0.9
